configure: make OpenLDAP test independent from X11
I missed that PostgreSQL can be build on MacOSX with OpenLDAP
support. Stephan had an incremental build failure because of it.

But instead of opting-in using test_openldap=yes on MacOSX, like
on Haiku, this drops the using_x11 dependency and defaults to
"yes", like it was before. This was more of a convenience then
reality anyway, and it makes really not much sense to disable
OpenLDAP in case of --without-x.

Change-Id: I959a3b30f1c369264174302f18267a4eb40a3a52